backend/requirements.txt
fastapi
uvicorn
openai
python-multipart
pytesseract
Pillow
numpy_financial
PyMuPDF
